Complete the collection of Card in two components

+++ depressive-back/src/main/java/com/example/depressive/mood/repository/UserCardsLogRepository.java
@@ -1,9 +1,7 @@
 package com.example.depressive.mood.repository;
 
 import com.example.depressive.mood.entity.UserCardsLog;
-import io.lettuce.core.dynamic.annotation.Param;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -11,7 +9,8 @@ import java.util.Optional;
 
 public interface UserCardsLogRepository extends JpaRepository<UserCardsLog, Long> {
     Optional<UserCardsLog> findByUserIdAndDate(Long userId, LocalDate date);
-    List<UserCardsLog> findByUserIdAndContentId(Long userId, Short contentId);
+    List<UserCardsLog> findByUserIdAndContentId(Long userId, Long contentId);
     Optional<UserCardsLog> findTopByUserIdAndContentIdOrderByDateDesc(Long userId, Short contentId);
     List<UserCardsLog> findByUserIdOrderByDateDesc(Long userId);
+
 }
